Created utilities.cc file
