+++ include/aspect/velocity_boundary_conditions/ascii_data.h
@@ -22,8 +22,8 @@
 #ifndef __aspect__velocity_boundary_conditions_ascii_data_h
 #define __aspect__velocity_boundary_conditions_ascii_data_h
 
+#include <aspect/velocity_boundary_conditions/interface.h>
 #include <aspect/simulator_access.h>
-
 #include <aspect/utilities.h>
